Avoid trying to lock on non-files
This is not supported on windows and makes little sense on POSIX

+++ lib/puppet/util/file_locking.rb
@@ -5,6 +5,7 @@ module Puppet::Util::FileLocking
 
     # Create a shared lock for reading
     def readlock(file)
+        raise ArgumentError, "%s is not a file" % file unless !File.exists?(file) or File.file?(file)
         Puppet::Util.sync(file).synchronize(Sync::SH) do
             File.open(file) { |f|
                 f.lock_shared { |lf| yield lf }
@@ -18,6 +19,7 @@ module Puppet::Util::FileLocking
         unless FileTest.directory?(File.dirname(file))
             raise Puppet::DevError, "Cannot create %s; directory %s does not exist" % [file, File.dirname(file)]
         end
+        raise ArgumentError, "%s is not a file" % file unless !File.exists?(file) or File.file?(file)
         tmpfile = file + ".tmp"
 
         unless mode
